+++ kurtosis-module/impl/grafana/grafana_launcher.go
@@ -1,15 +1,10 @@
 package grafana
 
 import (
-	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/service_launch_utils"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/static_files"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/services"
 	"github.com/kurtosis-tech/stacktrace"
-	"io"
-	"os"
-	"path"
-	"text/template"
 )
 
 const (
@@ -19,20 +14,16 @@ const (
 	httpPortId            = "http"
 	httpPortNumber uint16 = 3000
 
-	configDirectoriesPermission = 0755
+	datasourceConfigRelFilepath = "datasources/datasource.yml"
 
-	datasourcesConfigDirname = "datasources"
-	datasourceConfigFilename = "datasource.yml"
-
-	dashboardsConfigDirname          = "dashboards"
-	dashboardProvidersConfigFilename = "dashboard-providers.yml"
-	dashboardConfigFilename          = "dashboard.json"
+	// this is relative to the files artifact root
+	dashboardProvidersConfigRelFilepath = "dashboards/dashboard-providers.yml"
 
 	configDirpathEnvVar = "GF_PATHS_PROVISIONING"
 
-	grafanaConfigDirpathOnModule = "/tmp/grafana-config"
-
-	grafanaConfigDirpathOnService = "/config"
+	grafanaConfigDirpathOnService      = "/config"
+	grafanaDashboardsDirpathOnService  = "/dashboards"
+	grafanaDashboardsFilepathOnService = grafanaDashboardsDirpathOnService + "/dashboard.json"
 )
 
 var usedPorts = map[string]*services.PortSpec{
@@ -49,11 +40,11 @@ type dashboardProvidersConfigTemplateData struct {
 
 func LaunchGrafana(
 	enclaveCtx *enclaves.EnclaveContext,
-	datasourceConfigTemplate *template.Template,
-	dashboardProvidersConfigTemplate *template.Template,
+	datasourceConfigTemplate string,
+	dashboardProvidersConfigTemplate string,
 	prometheusPrivateUrl string,
 ) error {
-	artifactUuid, err := getGrafanaConfigDirArtifactUuid(
+	grafanaConfig, grafanaDashboards, err := getGrafanaConfigDirArtifactUuid(
 		enclaveCtx,
 		datasourceConfigTemplate,
 		dashboardProvidersConfigTemplate,
@@ -63,7 +54,7 @@ func LaunchGrafana(
 		return stacktrace.Propagate(err, "An error occurred getting the Grafana config directory files artifact")
 	}
 
-	containerConfigSupplier := getContainerConfigSupplier(artifactUuid)
+	containerConfigSupplier := getContainerConfigSupplier(grafanaConfig, grafanaDashboards)
 	_, err = enclaveCtx.AddService(serviceID, containerConfigSupplier)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred launching the grafana service")
@@ -79,111 +70,58 @@ func LaunchGrafana(
 // ====================================================================================================
 func getGrafanaConfigDirArtifactUuid(
 	enclaveCtx *enclaves.EnclaveContext,
-	datasourceConfigTemplate *template.Template,
-	dashboardProvidersConfigTemplate *template.Template,
+	datasourceConfigTemplate string,
+	dashboardProvidersConfigTemplate string,
 	prometheusPrivateUrl string,
-) (services.FilesArtifactUUID, error) {
-	datasourcesConfigDirpathOnModule := path.Join(grafanaConfigDirpathOnModule, datasourcesConfigDirname)
-	datasourceConfigFilepath := path.Join(datasourcesConfigDirpathOnModule, datasourceConfigFilename)
-	dashboardsConfigDirpathOnModule := path.Join(grafanaConfigDirpathOnModule, dashboardsConfigDirname)
-	dashboardProvidersConfigFilepath := path.Join(dashboardsConfigDirpathOnModule, dashboardProvidersConfigFilename)
-	dashboardConfigFilepath := path.Join(dashboardsConfigDirpathOnModule, dashboardConfigFilename)
+) (resultGrafanaConfig services.FilesArtifactUUID, resultGrafanaDashboards services.FilesArtifactUUID, resultErr error) {
 
-	dashboardConfigFilepathOnGrafanaContainer := path.Join(
-		grafanaConfigDirpathOnService,
-		dashboardsConfigDirname,
-		dashboardConfigFilename,
-	)
-
-	dirpathsToCreate := []string{
-		grafanaConfigDirpathOnModule,
-		datasourcesConfigDirpathOnModule,
-		dashboardsConfigDirpathOnModule,
-	}
-	for _, dirpath := range dirpathsToCreate {
-		if err := os.Mkdir(dirpath, configDirectoriesPermission); err != nil {
-			return "", stacktrace.Propagate(err, "An error occurred creating Grafana config directory '%v'", dirpathsToCreate)
-		}
-	}
-
-	datasourceTemplateData := datasourceConfigTemplateData{
+	datasourceData := datasourceConfigTemplateData{
 		PrometheusURL: prometheusPrivateUrl,
 	}
-	if err := service_launch_utils.FillTemplateToPath(datasourceConfigTemplate, datasourceTemplateData, datasourceConfigFilepath); err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred filling the datasource config template")
-	}
+	datasourceTemplateAndData := enclaves.NewTemplateAndData(datasourceConfigTemplate, datasourceData)
 
-	dashboardProvidersTemplateData := dashboardProvidersConfigTemplateData{
+	dashboardProvidersData := dashboardProvidersConfigTemplateData{
 		// Grafana needs to know where the dashboards config file will be on disk, which means we need to feed
 		//  it the *mounted* location on disk (on the Grafana container) when we generate this on the module container
-		DashboardsDirpath: dashboardConfigFilepathOnGrafanaContainer,
+		DashboardsDirpath: grafanaDashboardsFilepathOnService,
 	}
-	if err := service_launch_utils.FillTemplateToPath(dashboardProvidersConfigTemplate, dashboardProvidersTemplateData, dashboardProvidersConfigFilepath); err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred filling the dashboard config providers template")
-	}
+	dashboardProvidersTemplateAndData := enclaves.NewTemplateAndData(dashboardProvidersConfigTemplate, dashboardProvidersData)
 
-	if err := addGrafanaDashboardConfigToConfigDir(
-		static_files.GrafanaDashboardConfigFilepath,
-		dashboardConfigFilepath,
-	); err != nil {
-		return "", stacktrace.Propagate(
-			err,
-			"An error occurred copying Grafana dashboard config file '%v' to the Grafana config directory at '%v'",
-			static_files.GrafanaDashboardConfigFilepath,
-			dashboardConfigFilepath,
-		)
-	}
+	templateAndDataByDestRelFilepath := make(map[string]*enclaves.TemplateAndData)
+	templateAndDataByDestRelFilepath[datasourceConfigRelFilepath] = datasourceTemplateAndData
+	templateAndDataByDestRelFilepath[dashboardProvidersConfigRelFilepath] = dashboardProvidersTemplateAndData
 
-	artifactUuid, err := enclaveCtx.UploadFiles(grafanaConfigDirpathOnModule)
+	grafanaConfig, err := enclaveCtx.RenderTemplates(templateAndDataByDestRelFilepath)
 	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred uploading Grafana config dir at '%v'", grafanaConfigDirpathOnModule)
+		return "", "", stacktrace.Propagate(err, "An error occurred rendering Grafana templates")
 	}
 
-	return artifactUuid, nil
+	grafanaDashboards, err := enclaveCtx.UploadFiles(static_files.GrafanaDashboardsConfigDirpath)
+	if err != nil {
+		return "", "", stacktrace.Propagate(err, "An error occurred uploading Grafana dashboard.json")
+	}
+
+	return grafanaConfig, grafanaDashboards, nil
 }
 
 func getContainerConfigSupplier(
-	configDirArtifactUuid services.FilesArtifactUUID,
+	grafanaConfig services.FilesArtifactUUID,
+	grafanaDashboards services.FilesArtifactUUID,
 ) func(privateIpAddr string) (*services.ContainerConfig, error) {
-	// We need the path.Base() here because Kurtosis doesn't flatten directories yet
 	containerConfigSupplier := func(privateIpAddr string) (*services.ContainerConfig, error) {
 		containerConfig := services.NewContainerConfigBuilder(
 			imageName,
 		).WithUsedPorts(
 			usedPorts,
 		).WithEnvironmentVariableOverrides(map[string]string{
-			configDirpathEnvVar: grafanaConfigDirpathOnModule,
+			configDirpathEnvVar: grafanaConfigDirpathOnService,
 		}).WithFiles(map[services.FilesArtifactUUID]string{
-			configDirArtifactUuid: grafanaConfigDirpathOnService,
+			grafanaConfig:     grafanaConfigDirpathOnService,
+			grafanaDashboards: grafanaDashboardsDirpathOnService,
 		}).Build()
 
 		return containerConfig, nil
 	}
 
 	return containerConfigSupplier
-}
-
-func addGrafanaDashboardConfigToConfigDir(srcFilepath, destFilepath string) error {
-	// Copy the config file from the static files
-	srcFp, err := os.Open(srcFilepath)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred opening Grafana dashboard config file '%v'", srcFilepath)
-	}
-	defer srcFp.Close()
-
-	destFp, err := os.Create(destFilepath)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating dashboard config file '%v'", destFilepath)
-	}
-	defer destFp.Close()
-
-	if _, err := io.Copy(destFp, srcFp); err != nil {
-		return stacktrace.Propagate(
-			err,
-			"An error occurred copying bytes from dashboard config source file '%v' to destination file '%v'",
-			srcFilepath,
-			destFilepath,
-		)
-	}
-	return nil
 }