+++ kurtosis-module/impl/static_files/static_files.go
@@ -1,11 +1,5 @@
 package static_files
 
-import (
-	"github.com/kurtosis-tech/stacktrace"
-	"path"
-	"text/template"
-)
-
 const (
 	// The path on the module container where static files are housed
 	staticFilesDirpath = "/static-files"
@@ -21,7 +15,7 @@ const (
 	CLGenesisGenerationMnemonicsTemplateFilepath = clGenesisGenerationConfigDirpath + "/mnemonics.yaml.tmpl"
 
 	// Prefunded keys
-	prefundedKeysDirpath = staticFilesDirpath + "/genesis-prefunded-keys"
+	prefundedKeysDirpath     = staticFilesDirpath + "/genesis-prefunded-keys"
 	GethPrefundedKeysDirpath = prefundedKeysDirpath + "/geth"
 
 	// Forkmon config
@@ -32,22 +26,7 @@ const (
 
 	//Grafana config
 	grafanaConfigDirpath                            = "/grafana-config"
-	GrafanaDatasourceConfigTemplateFilepath         = staticFilesDirpath + grafanaConfigDirpath + "/datasource.yml.tmpl"
-	GrafanaDashboardsConfigDirpath                  = staticFilesDirpath + grafanaConfigDirpath + "/dashboards"
-	GrafanaDashboardProvidersConfigTemplateFilepath = GrafanaDashboardsConfigDirpath + "/dashboard-providers.yml.tmpl"
-	GrafanaDashboardConfigFilepath                  = GrafanaDashboardsConfigDirpath + "/dashboard.json"
+	GrafanaDatasourceConfigTemplateFilepath         = staticFilesDirpath + grafanaConfigDirpath + "/templates/datasource.yml.tmpl"
+	GrafanaDashboardProvidersConfigTemplateFilepath = staticFilesDirpath + grafanaConfigDirpath + "/templates/dashboard-providers.yml.tmpl"
+	GrafanaDashboardsConfigDirpath                  = staticFilesDirpath + grafanaConfigDirpath + "/dashboards/dashboard.json"
 )
-
-func ParseTemplate(filepath string) (*template.Template, error) {
-	tmpl, err := template.New(
-		// For some reason, the template name has to match the basename of the file:
-		//  https://stackoverflow.com/questions/49043292/error-template-is-an-incomplete-or-empty-template
-		path.Base(filepath),
-	).ParseFiles(
-		filepath,
-	)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred parsing template file '%v'", filepath)
-	}
-	return tmpl, nil
-}